+++ layouts/error.vue
@@ -1,16 +1,22 @@
 <template>
-  <div class="hero">
-    <div class="hero-body">
-      <div class="container has-text-centered _card-input">
-        <p class="title is-3" v-if="error.statusCode === 404">
-          {{ $t('error.page_not_found') }}
-        </p>
-        <p class="title is-3" v-else>{{ error.message }}</p>
-        <nuxt-link class="subtitle is-4 has-text-link" :to="localePath('/')"
-          >{{ $t('error.home_link') }}</nuxt-link
-        >
-      </div>
-    </div>
+  <div class="_center_view">
+    <b-message
+      type="is-danger"
+      has-icon
+      size="is-large"
+      v-if="error.statusCode === 404"
+    >
+      {{ $t("error.page_not_found") }}
+      <br />
+      <nuxt-link :to="localePath('/')">{{ $t("error.home_link") }}</nuxt-link>
+    </b-message>
+    <b-message type="is-danger" has-icon size="is-large" v-else>
+      {{ error.message }}
+      <br />
+      <nuxt-link :to="localePath('/')">{{
+        $t("error.home_link")
+      }}</nuxt-link>
+    </b-message>
   </div>
 </template>
 
@@ -19,8 +25,8 @@ export default {
   props: ["error"],
   head() {
     return {
-      title: this.$t('error.head.title')
-    }
+      title: this.$t("error.head.title")
+    };
   }
 };
 </script>